03 Articles 2.0 Exercise
Articles 2.0
Change the program in such a way, that you will be able to store a list of articles. You will not need to use the previous methods anymore (except the "ToString()"). On the first line, you will receive the number of articles. On the next lines, you will receive the articles in the same format as in the previous problem: "{title}, {content}, {author}". Finally, you will receive a string: "title", "content" or an "author". You need to order the articles alphabetically, based on the given criteria.